+++ app/io/pathfinder/controllers/VehicleController.scala
@@ -1,0 +1,5 @@
+package io.pathfinder.controllers
+
+import io.pathfinder.models.Vehicle
+
+class VehicleController extends CrudController[Long,Vehicle](Vehicle)
