+++ app/io/pathfinder/controllers/ViewController.scala
@@ -1,0 +1,9 @@
+package io.pathfinder.controllers
+
+import play.api.mvc.{Controller,Action}
+
+class ViewController extends Controller {
+    def index = Action {
+        Ok("Pathfinder Webservice")
+    }
+}
